Update readme.md (#10)
Explain that it's only necessary to build gonk, not gecko or gaia. Explain simple patch and reference the git diff in pastebin.
Add dependencies for bindgen which don't appear to be included in Ubuntu 18.04, servo or B2G dependencies.
State explicitly that the device must already be flashed with B2G before flashing and running servonk.
Add that Nexus-5-L is also tested and works.